moves bundles loading methods into its own file
