+++ modules/@apostrophecms/template/index.js
@@ -119,6 +119,7 @@ module.exports = {
   },
   methods(self) {
     return {
+      ...require('./lib/bundlesLoader')(self),
 
       // Add helpers in the namespace for a particular module.
       // They will be visible in nunjucks at
@@ -698,7 +699,7 @@ module.exports = {
         try {
           const content = await module.render(req, template, args);
 
-          const filledContent = addExtraBundles({
+          const filledContent = self.insertBundlesMarkup({
             page: req.data.bestPage,
             scene,
             template,
@@ -735,157 +736,6 @@ module.exports = {
           } else {
             return url.replace('&aposRefresh=1', '');
           }
-        }
-
-        function addExtraBundles ({
-          page = {},
-          template = '',
-          scene,
-          content,
-          scriptsPlaceholder,
-          stylesheetsPlaceholder,
-          widgetsBundles = {}
-        }) {
-          if (!scriptsPlaceholder && !stylesheetsPlaceholder) {
-            return content;
-          }
-
-          const { es5 } = self.apos.asset.options;
-          const { extraBundles } = self.apos.asset;
-          const renderMarkup = renderBundleMarkup(self.apos.asset.getAssetBaseUrl());
-          const jsMainBundle = renderMarkup({
-            fileName: `${scene}-module-bundle`,
-            fileNameNoMod: `${scene}-nomodule-bundle`,
-            ext: 'js',
-            es5
-          });
-          const cssMainBundle = renderMarkup({
-            fileName: `${scene}-bundle`,
-            ext: 'css'
-          });
-
-          if (scene === 'apos') {
-            return loadAllBundles({
-              content,
-              scriptsPlaceholder,
-              stylesheetsPlaceholder,
-              extraBundles,
-              renderMarkup,
-              jsMainBundle,
-              cssMainBundle,
-              es5
-            });
-          }
-
-          const templateType = template.substring(template.lastIndexOf(':') + 1);
-          const { webpack } = page.type ? self.apos.modules[page.type].__meta : {};
-
-          const configs = Object.values(webpack || {}).reduce((acc, config) => ({
-            ...acc,
-            ...config && config.bundles
-          }), widgetsBundles);
-
-          const { jsBundles, cssBundles } = Object.entries(configs)
-            .reduce((acc, [ name, { templates } ]) => {
-              if (templates && !templates.includes(templateType)) {
-                return acc;
-              }
-
-              const jsMarkup = scriptsPlaceholder &&
-              extraBundles.js.includes(`${name}.js`) &&
-                renderMarkup({
-                  fileName: name,
-                  ext: 'js'
-                });
-
-              const cssMarkup = stylesheetsPlaceholder &&
-                extraBundles.css.includes(`${name}.css`) &&
-                renderMarkup({
-                  fileName: name,
-                  ext: 'css'
-                });
-
-              return {
-                jsBundles: stripIndent`
-                  ${acc.jsBundles}
-                  ${jsMarkup || ''}
-                `,
-                cssBundles: stripIndent`
-                  ${acc.cssBundles}
-                  ${cssMarkup || ''}
-                `
-              };
-            }, {
-              jsBundles: jsMainBundle,
-              cssBundles: cssMainBundle
-            });
-
-          return content
-            .replace(scriptsPlaceholder, jsBundles)
-            .replace(stylesheetsPlaceholder, cssBundles);
-        }
-
-        function loadAllBundles({
-          content,
-          extraBundles,
-          scriptsPlaceholder,
-          stylesheetsPlaceholder,
-          jsMainBundle,
-          cssMainBundle,
-          renderMarkup
-        }) {
-          const jsBundles = extraBundles.js.reduce((acc, bundle) => {
-            const bundleMarkup = renderMarkup({
-              fileName: bundle.replace('.js', ''),
-              ext: 'js'
-            });
-
-            return stripIndent`
-              ${acc}
-              ${bundleMarkup}
-            `;
-
-          }, jsMainBundle);
-
-          const cssBundles = extraBundles.css.reduce((acc, bundle) => {
-            const bundleMarkup = renderMarkup({
-              fileName: bundle.replace('.css', ''),
-              ext: 'css'
-            });
-
-            return stripIndent`
-              ${acc}
-              ${bundleMarkup}
-            `;
-
-          }, cssMainBundle);
-
-          return content
-            .replace(scriptsPlaceholder, jsBundles)
-            .replace(stylesheetsPlaceholder, cssBundles);
-        }
-
-        function renderBundleMarkup (base) {
-          return ({
-            fileName, fileNameNoMod, ext = 'js', es5 = false
-          }) => {
-            if (ext === 'css') {
-              return self.apos.template.safe(stripIndent`
-                <link href="${base}/${fileName}.css" rel="stylesheet" />
-              `);
-            }
-
-            if (es5) {
-              return self.apos.template.safe(stripIndent`
-                <script nomodule src="${base}/${fileNameNoMod}.${ext}"></script>
-                <script type="module" src="${base}/${fileName}.${ext}"></script>
-              `);
-            }
-
-            return self.apos.template.safe(stripIndent`
-              <script src="${base}/${fileName}.${ext}"></script>
-            `);
-          };
         }
       },
